+++ fe/fe-core/src/main/java/com/starrocks/lake/StarOSAgent.java
@@ -1,6 +1,6 @@
 // This file is licensed under the Elastic License 2.0. Copyright 2021-present, StarRocks Limited.
 
-package com.starrocks.catalog;
+package com.starrocks.lake;
 
 import com.google.common.base.Preconditions;
 import com.google.common.collect.Lists;
