+++ fe/fe-core/src/test/java/com/starrocks/common/proc/LakeTabletsProcNodeTest.java
@@ -14,7 +14,7 @@ import com.starrocks.catalog.MaterializedIndex;
 import com.starrocks.catalog.Partition;
 import com.starrocks.catalog.PartitionInfo;
 import com.starrocks.catalog.SinglePartitionInfo;
-import com.starrocks.catalog.StarOSAgent;
+import com.starrocks.lake.StarOSAgent;
 import com.starrocks.catalog.Tablet;
 import com.starrocks.catalog.TabletMeta;
 import com.starrocks.catalog.Type;
