+++ fe/fe-core/src/test/java/com/starrocks/lake/ShardDeleterTest.java
@@ -2,7 +2,6 @@
 
 package com.starrocks.lake;
 
-import com.starrocks.catalog.StarOSAgent;
 import com.starrocks.common.jmockit.Deencapsulation;
 import com.starrocks.lake.proto.DropTabletRequest;
 import com.starrocks.lake.proto.DropTabletResponse;
@@ -32,7 +31,7 @@ public class ShardDeleterTest {
     private GlobalStateMgr globalStateMgr;
 
     @Mocked
-    private SystemInfoService  systemInfoService;
+    private SystemInfoService systemInfoService;
 
     @Mocked
     private StarOSAgent starOSAgent;
@@ -92,7 +91,7 @@ public class ShardDeleterTest {
 
         new Expectations() {
             {
-                lakeService.dropTablet((DropTabletRequest)any).get();
+                lakeService.dropTablet((DropTabletRequest) any).get();
                 minTimes = 0;
                 result = response;
